requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
cycler==0.10.0
discord.py==1.3.2
idna==2.9
kiwisolver==1.1.0
matplotlib==3.2.1
multidict==4.7.5
numpy==1.18.2
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
websockets==8.1
yarl==1.4.2
